restructured project, tried out relative submodules (still WIP)
