+++ src/chessai/__init__.py
@@ -1,0 +1,3 @@
+
+import .dataset import *
+import .pre_train import *
