+++ src/chessai/dataset/gm_games_dataset.py
