+++ src/chessai/legacy/keras-test.py
