+++ src/chessai/pre_train/README.md
