+++ src/chessai/pre_train/pretrain_rating.py
@@ -7,20 +7,20 @@ import tensorflow as tf
 from tensorflow import keras
 import chesslib
 
-from dataset import ChessGmGamesDataset
+from ..dataset import ChessGmGamesDataset
 from . import ChessRatingModel
 
 
-class DeepQTrainingSession(object):
+class RatingTrainingSession(object):
 
     def __init__(self, params: dict):
 
-        super(DeepQTrainingSession, self).__init__()
+        super(RatingTrainingSession, self).__init__()
         self.params = params
 
         # create training and evaluation datasets
-        gmGamesDataset = ChessGmGamesDataset()
-        self.train_dataset, self.eval_dataset = gmGamesDataset.load_datasets(params['batch_size'])
+        dataset = ChessGmGamesDataset(params['batch_size'])
+        self.train_dataset, self.eval_dataset = dataset.load_datasets()
 
         # create model to be trained
         self.model = ChessRatingModel(params)
@@ -42,7 +42,7 @@ class DeepQTrainingSession(object):
         # create model checkpoints
         self.checkpoint = tf.train.Checkpoint(optimizer=self.optimizer, net=self.model)
         self.manager = tf.train.CheckpointManager(self.checkpoint, './models/pretrain', max_to_keep=5)
-        # TODO: load pretrained feature extractor and ratings here ...
+        # TODO: load pretrained feature extractor here ...
 
         # create logging metrics
         self.train_loss = tf.keras.metrics.Mean(name="train_loss")
@@ -149,7 +149,7 @@ def main():
         'total_train_batches': 2774,
     }
 
-    session = DeepQTrainingSession(params)
+    session = RatingTrainingSession(params)
     session.run_training()
